user authentication in all to do requests

+++ routers/todos.py
@@ -6,7 +6,7 @@ import models
 from models import Todos
 from database import engine, SessionLocal
 from starlette import status
-from routers import auth
+from .auth import get_current_user
 
 router = APIRouter()
 
@@ -18,6 +18,7 @@ def get_db():
         db.close()
 
 db_dependency = Annotated[Session, Depends(get_db)]
+user_dependency = Annotated[dict, Depends(get_current_user)]
 
 class TodoRequest(BaseModel):
     title: str =Field(min_length=3, max_length=50)
@@ -28,27 +29,39 @@ class TodoRequest(BaseModel):
 
 
 @router.get("/")
-async def read_all(db: db_dependency):
-    return db.query(Todos).all()
+async def read_all(user:user_dependency,db: db_dependency):
+    if user is None:
+        return HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Invalid authentication credentials")
+    return db.query(Todos).filter(Todos.owner_id == user.get("id")).all()
 
 
 @router.get("/todo/{todo_id}", status_code=status.HTTP_200_OK)
-async def read_todo(db: db_dependency, todo_id: int=Path(gt=0)):
-    todo_model= db.query(Todos).filter(Todos.id == todo_id).first()
+async def read_todo(user:user_dependency,db: db_dependency, todo_id: int=Path(gt=0)):
+    if user is None:
+        return HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Invalid authentication credentials")
+    todo_model= db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("id")).first()
     if todo_model is None:
         return HTTPException(status_code=404, detail="Todo not found")
     return todo_model
 
 
 @router.post("/todo", status_code=status.HTTP_201_CREATED)
-async def create_todo(db: db_dependency, todo_request: TodoRequest):
-    todo_model = Todos(title=todo_request.title, description=todo_request.description, priority=todo_request.priority, completed=todo_request.completed)
+async def create_todo(user:user_dependency,db: db_dependency, todo_request: TodoRequest):
+    if user is None:
+        return HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Invalid authentication credentials")
+    todo_model = Todos(title=todo_request.title,
+                    description=todo_request.description,
+                    priority=todo_request.priority,
+                    completed=todo_request.completed,
+                    owner_id=user.get("id"))
     db.add(todo_model)
     db.commit()
 
 @router.put("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db: db_dependency, todo_request: TodoRequest, todo_id: int=Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def update_todo(user:user_dependency,db: db_dependency, todo_request: TodoRequest, todo_id: int=Path(gt=0)):
+    if user is None:
+        return HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Invalid authentication credentials")
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("id")).first()
     if todo_model is None:
         return HTTPException(status_code=404, detail="Todo not found")
     todo_model.title = todo_request.title
@@ -60,9 +73,11 @@ async def update_todo(db: db_dependency, todo_request: TodoRequest, todo_id: int
 
 
 @router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db:db_dependency,todo_id:int=Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def delete_todo(user:user_dependency,db:db_dependency,todo_id:int=Path(gt=0)):
+    if user is None:
+        return HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Invalid authentication credentials")
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("id")).first()
     if todo_model is None:
         return HTTPException(status_code=404, detail="Todo not found")
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("id")).delete()
     db.commit()
